Resetting console mocks
I noticed this bug when a fellow had passing code for the 4th debug problem but failing code for the 3rd debug problem. The issue is that the call count for the mock console wasn't being restored if the test failed. This solution does work but using a global `logSpy` and reassigning it before each test feels wrong to me but I couldn't find good examples of how to do this a "proper" way online. I also looked at this thread below which we could use instead (`jest.clearAllMocks()` in an `afterEach`)

https://stackoverflow.com/questions/47812801/how-to-reset-jest-mock-functions-calls-count-before-every-test